- Finalise the database cleanup functionality. It's UNTESTED! I'll run this in my homelab for a while when I get a chance and see how it goes.

+++ src/functions/database_functions.py
@@ -7,18 +7,18 @@ CLIENT_CONNECTION_STRING = pymongo.MongoClient("mongodb://mongodb:27017/")
 LIST_EXISTING_DBS = CLIENT_CONNECTION_STRING.list_database_names()
 
 
-# Return a list of existing databases in our mongoDB instance
+# Return a list of existing databases in our mongoDB instance.
 def list_existing_databases():
     return LIST_EXISTING_DBS
 
 
-# Return a list of existing collections in a database
+# Return a list of existing collections in a database.
 def list_existing_collections(database):
     DATABASE = CLIENT_CONNECTION_STRING[database]
     return DATABASE.list_collection_names()
 
 
-# Insert a document to a mongoDB collection
+# Insert a document to a mongoDB collection.
 def insert_document_in_mongodb(database, collection, dict):
     DATABASE = CLIENT_CONNECTION_STRING[database]
     COL = DATABASE[collection]
@@ -30,7 +30,7 @@ def insert_document_in_mongodb(database, collection, dict):
 
 
 # Check if a document exists in our mongoDB collection (based on the date/time the document was written to our
-# collection)
+# collection).
 def check_if_document_exists_in_mongodb(database, collection, dict):
     DATABASE = CLIENT_CONNECTION_STRING[database]
     COL = DATABASE[collection]
@@ -43,7 +43,7 @@ def check_if_document_exists_in_mongodb(database, collection, dict):
             return False
 
 
-# Get the biggest winning position for a certain time period
+# Get the biggest winning position for a certain time period.
 # Use MongoDB queries for this and minimal python code
 # Try something like this https://www.mongodb.com/community/forums/t/fetch-data-with-max-and-between-condition/3973
 def get_biggest_winning_position(database, collection, time_period):
@@ -57,7 +57,7 @@ def get_losing_winning_position(database, collection, time_period):
     print("working on it!")
 
 
-# Delete a document from a collection in MongoDB based on its updated_time being greater than time_limit_days
+# Delete a document from a collection in MongoDB based on its updated_time being greater than time_limit_days.
 def delete_document_from_mongodb(database, collection, time_limit_days):
     DATABASE = CLIENT_CONNECTION_STRING[database]
     COL = DATABASE[collection]
@@ -75,11 +75,12 @@ def delete_document_from_mongodb(database, collection, time_limit_days):
         logging.error("ERROR:" + " " + "Failed to delete document from collection" + " " + collection)
 
 
-# GENERATED WITH GITHUB CO-PILOT! Utilising our list_existing_databases function, define a function to loop through
-# the list of existing databases that it returns and for each collection within our databases, utilise our
-# delete_document_from_mongodb function to loop through each document within the collection and delete those that are
-# older than one year old.
+# Utilising our list_existing_databases function, define a function to loop through the list of existing databases
+# that it returns and for each collection within our databases that are NOT the "admin", "config" or "local"
+# databases, utilise our delete_document_from_mongodb function to loop through each document within the collection
+# and delete those documents that have an updated_time greater than the time_limit_days.
 def clean_up_mongodb(time_limit_days):
     for database in list_existing_databases():
-        for collection in list_existing_collections(database):
-            delete_document_from_mongodb(database, collection, time_limit_days)
+        if not database == "admin" or not database == "config" or not database == "local":
+            for collection in list_existing_collections(database):
+                delete_document_from_mongodb(database, collection, time_limit_days)
